#1:: delimiter for item parts update
previously was using semicolons for item_parts__ip field but it was too common in the field. Note: The functional delimiter in the script is actually a line break ("\n") not necessarily the #1:: delimiter. That also means that line breaks in a field such as Container/Item Annotations would not work with this script. Line breaks can only be used after each "#1::" entries.